fixed the positioning of the tiles
we now take user _screen_ coordinates of the tile in the tilset file and translate
it to 0;0 in our coordinate system using these coordinates.
Previously we used getBBox() for that purpose but that function doesn't take translations
into account so all the elements that were translated in the tileset file ended up in the wrong place.
Now we use getScreenBBox() which is specified in the SVG 1.2 standard but seems to be present ATM
only in the Opera so we have to use JS method elsewhere.