set the svg:viewBox attribute after laying out the board
Under some browsers laying out things doesn't work quite as
I would expect if the viewBox is already set. So this is basically
just a workaround for such cases.

We also now use predefined size for the focusbox size since using the
rendered tile size gives us smaller height & width than it really is and
reading the sizes from the original tilesets gives larger size than it
really is. This seems to be caused by the viewBox in some way, as if the
viewBox scale factor isn't taken into account when calculating getScreenBBox()
value