allow recursive hiding of the original dom tree after the board has been rendered
Under some browsers hiding the original DOM before we ask for screen bbox returns seemingly
random results which is correct in a way that since it's not on the screen it obviously
can't have a bbox.
This isn't a pretty solution of course since the user sees the rendering artifacts for a moment.
Better solution would probably be to move the original tileset into part of the canvas that
isn't visible.